support multi-group log commands
With this change, the first argument to the command is parsed as a
regular expression and matched against the list of available log groups.
For each matched log group, the relevant command is executed in
parallel.

The following commands support this pattern:

* `get`
* `streams`
* `watch`

The log group name is added to the formatted log output, to distinguish
logs from different log groups.

If you want to get an idential result as what you got before this
commit, you would have to wrap your group name in regular expression
begin and end tokens, meaning this:

    saw stream "my-log-group"

Must now be written as:

    saw stream "^my-log-group$"

Signed-off-by: Jean Mertz <jean@mertz.fm>